ProductBillingPeriod<P> type
