+++ handlers/press-reader-entitlements/src/supporterProductData.ts
@@ -2,8 +2,8 @@ import { sortBy } from '@modules/arrayFunctions';
 import type {
 	ProductCatalog,
 	ProductKey,
-} from '@modules/product-catalog/newProductCatalogTypes';
-import { ProductCatalogHelper } from '@modules/product-catalog/newProductCatalogTypes';
+} from '@modules/product-catalog/productCatalog';
+import { ProductCatalogHelper } from '@modules/product-catalog/productCatalog';
 import type { Stage } from '@modules/stage';
 import type { SupporterRatePlanItem } from '@modules/supporter-product-data/supporterProductData';
 import { getSupporterProductData } from '@modules/supporter-product-data/supporterProductData';
