+++ handlers/product-switch-api/test/contributionToSupporterPlus.test.ts
@@ -5,7 +5,7 @@
 import type { EmailMessageWithUserId } from '@modules/email/email';
 import { ValidationError } from '@modules/errors';
 import { generateProductCatalog } from '@modules/product-catalog/generateProductCatalog';
-import type { ProductCatalog } from '@modules/product-catalog/newProductCatalogTypes';
+import type { ProductCatalog } from '@modules/product-catalog/productCatalog';
 import type { ZuoraSubscription } from '@modules/zuora/zuoraSchemas';
 import {
 	zuoraAccountSchema,
