+++ handlers/update-supporter-plus-amount/src/sendEmail.ts
@@ -1,7 +1,7 @@
 import type { EmailMessageWithUserId } from '@modules/email/email';
 import { DataExtensionNames } from '@modules/email/email';
 import type { Currency } from '@modules/internationalisation/currency';
-import type { ProductBillingPeriod } from '@modules/product-catalog/newProductCatalogTypes';
+import type { ProductBillingPeriod } from '@modules/product-catalog/productBillingPeriods';
 import type dayjs from 'dayjs';
 
 export type EmailFields = {
