+++ handlers/update-supporter-plus-amount/src/updateSupporterPlusAmount.ts
@@ -4,11 +4,11 @@ import type { Currency } from '@modules/internationalisation/currency';
 import { isSupportedCurrency } from '@modules/internationalisation/currency';
 import { getIfDefined } from '@modules/nullAndUndefined';
 import { prettyPrint } from '@modules/prettyPrint';
+import type { ProductBillingPeriod } from '@modules/product-catalog/productBillingPeriods';
 import type {
-	ProductBillingPeriod,
 	ProductCatalog,
 	ProductRatePlan,
-} from '@modules/product-catalog/newProductCatalogTypes';
+} from '@modules/product-catalog/productCatalog';
 import { zuoraDateFormat } from '@modules/zuora/common';
 import { getAccount } from '@modules/zuora/getAccount';
 import { getSubscription } from '@modules/zuora/getSubscription';
