+++ handlers/update-supporter-plus-amount/test/updateSupporterPlusAmounts.test.ts
@@ -1,5 +1,5 @@
 import { generateProductCatalog } from '@modules/product-catalog/generateProductCatalog';
-import type { ProductCatalog } from '@modules/product-catalog/newProductCatalogTypes';
+import type { ProductCatalog } from '@modules/product-catalog/productCatalog';
 import { Logger } from '@modules/zuora/logger';
 import { zuoraSubscriptionResponseSchema } from '@modules/zuora/zuoraSchemas';
 import zuoraCatalogFixture from '../../../modules/zuora-catalog/test/fixtures/catalog-code.json';
