+++ handlers/user-benefits/test/benefitsIdentityId.test.ts
@@ -1,5 +1,5 @@
 import { getUserBenefitsExcludingStaff } from '@modules/product-benefits/userBenefits';
-import type { ProductCatalogHelper } from '@modules/product-catalog/newProductCatalogTypes';
+import type { ProductCatalogHelper } from '@modules/product-catalog/productCatalog';
 import type { Stage } from '@modules/stage';
 import type { APIGatewayProxyEvent } from 'aws-lambda';
 import { benefitsIdentityIdHandler } from '../src/benefitsIdentityId';
