+++ modules/product-catalog/README.md
@@ -40,12 +40,12 @@ const catalog = await getProductCatalogFromApi('PROD');
 const contributionChargeId = catalog.SupporterPlus.ratePlans.Monthly.charges.Contribution.id;
 ```
 ## Implementation
-The mapping between our object model and Zuora's catalog is defined in the files `generateProductCatalog.ts`, `zuoraToProductNameMapping.ts` and `generateTypeObject.ts`.
+The mapping between our object model and Zuora's catalog is defined in the files `generateProductCatalog.ts`, `zuoraToProductNameMapping.ts` and `generateProductBillingPeriods.ts`.
 
-These files contain functions to generate the product catalog json from the Zuora catalog and also to create a type object which we use to define the types of which the catalog is comprised. This type object is written to `typeObject.ts` and checked into the repo. If it ever needs to be regenerated, it can be done so by running 
+These files contain functions to generate the product catalog json from the Zuora catalog and also to create a Zod schema for this json which we then use to infer the types of which the catalog is comprised. This schema is written to `productCatalogSchema.ts` and checked into the repo. If it ever needs to be regenerated, it can be done so by running 
 ```shell
 # requires fresh Janus credentials
-pnpm --filter product-catalog generateTypes
+pnpm --filter product-catalog generateSchema
 ```
 from the root of the repository.
