+++ modules/product-catalog/package.json
@@ -2,8 +2,7 @@
   "name": "product-catalog",
   "version": "1.0.0",
   "scripts": {
-    "generateTypes": "ts-node src/generateTypeObjectCommand.ts && prettier --write src/typeObject.ts",
-    "generateSchema": "ts-node src/generateProductSchemaCommand.ts && prettier --write src/productCatalogSchema.ts",
+    "generateSchema": "ts-node src/generateSchemaCommand.ts && prettier --write src/productCatalogSchema.ts && prettier --write src/productBillingPeriods.ts",
     "updateSnapshots": "jest -u --group=-integration",
     "test": "jest --group=-integration",
     "it-test": "jest --group=integration",
