+++ modules/product-catalog/src/generateSchemaCommand.ts
@@ -1,14 +1,14 @@
 import fs from 'fs';
 import { getZuoraCatalogFromS3 } from '@modules/zuora-catalog/S3';
-import { generateSchema } from '@modules/product-catalog/generateProductCatalogSchema';
+import { generateProductBillingPeriods } from '@modules/product-catalog/generateProductBillingPeriods';
+import { generateSchema } from '@modules/product-catalog/generateSchema';
 
 const writeSchemaToFile = async () => {
 	const prodCatalog = await getZuoraCatalogFromS3('PROD');
-	// const types = generateTypeObjects(prodCatalog);
-	// const activeTypesString = JSON.stringify(types.active, null, 2);
-	// const inactiveTypesString = JSON.stringify(types.inactive, null, 2);
 	const generatedSchema = generateSchema(prodCatalog);
 	fs.writeFileSync('./src/productCatalogSchema.ts', generatedSchema);
+	const productBillingPeriods = generateProductBillingPeriods(prodCatalog);
+	fs.writeFileSync('./src/productBillingPeriods.ts', productBillingPeriods);
 };
 
 void (async function () {
