+++ modules/product-catalog/test/productCatalogMapping.test.ts
@@ -2,7 +2,7 @@ import { findDuplicates } from '@modules/arrayFunctions';
 import type { Stage } from '@modules/stage';
 import { ZuoraCatalogHelper } from '@modules/zuora-catalog/zuoraCatalog';
 import { generateProductCatalog } from '@modules/product-catalog/generateProductCatalog';
-import { ProductCatalogHelper } from '@modules/product-catalog/newProductCatalogTypes';
+import { ProductCatalogHelper } from '@modules/product-catalog/productCatalog';
 import codeZuoraCatalog from '../../zuora-catalog/test/fixtures/catalog-code.json';
 import prodZuoraCatalog from '../../zuora-catalog/test/fixtures/catalog-prod.json';
